requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.1
h11==0.13.0
idna==3.3
joblib==1.1.0
kanjize==1.0.0
numpy==1.22.3
outcome==1.1.0
Pillow==9.1.0
py-ogp-parser==0.2.1
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SudachiDict-core==20211220
SudachiPy==0.6.3
threadpoolctl==3.1.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
